feat: restructure XP management as /leveling xp|level subcommand groups
