+++ apps/bot/src/commands.ts
@@ -49,7 +49,7 @@ import { shopCommand } from './commands/economy/shop.js';
 import { spawnItemCommand } from './commands/economy/spawnItem.js';
 import { useCommand } from './commands/economy/use.js';
 import { workCommand } from './commands/economy/work.js';
-import { xpCommand } from './commands/admin/xp.js';
+import { levelingCommand } from './commands/admin/leveling.js';
 import { serverstatsCommand } from './commands/utility/serverstats.js';
 import { setupCommand } from './commands/admin/setup.js';
 import { statsCommand } from './commands/utility/stats.js';
@@ -184,7 +184,7 @@ export const commands: SlashCommandDefinition[] = [
   pulseCommand,
   evaluationsCommand,
   rappelCommand,
-  xpCommand,
+  levelingCommand,
 ];
 
 /**
